feat(mcp): compute loopover_lint_pr_text in-process from the shared engine
The local MCP server's loopover_lint_pr_text proxied over HTTP
(apiPost("/v1/lint/pr-text", …)) because buildPrTextLint lived in the
host-bound signals/engine.ts, which is excluded from @loopover/engine's tsc
emit and so never reached the published dist/.

Extract buildPrTextLint and its types into a self-contained, buildable
signals/pr-text-lint.ts that reuses the already-extracted shared rubric
(GENERIC_COMMIT_PATTERN/hasClearNoIssueRationale from slop.ts, tokenize from
predicted-gate-engine.ts) so there is one definition, not a mirror.
signals/engine.ts re-exports it, keeping the API route and remote MCP server
unchanged, and the engine barrel now exposes it.

Bump loopover-mcp's @loopover/engine dependency from ^1.0.0 (which resolved
the stale registry 1.0.0 predating this export) to ^3.0.0 so it links the
workspace engine like loopover-miner already does, then call buildPrTextLint
directly in the tool handler. A user running the local MCP server can now lint
PR text fully offline, matching the remote server's in-process behavior.

Closes #6268
